+++ src/integrator_whfast.c
@@ -215,7 +215,7 @@ WIN32DLL_API ErrorStatus whfast(
 
     bool is_output = (output_param->method != OUTPUT_METHOD_DISABLED);
     int *__restrict output_count_ptr = &(output_param->output_count_);
-    const int output_interval = output_param->output_interval;
+    const double output_interval = output_param->output_interval;
     double next_output_time = output_interval;
 
     double *__restrict t_ptr = &(simulation_status->t);
